follow_me_laptop_test/requirements.txt
# Follow Me — Dependencies
#
# ── Laptop / cross-platform ──────────────────────────────────────────────────
# Cài đặt:
#   pip install -r requirements.txt
#
# Lưu ý: ultralytics tự tải yolov8n.pt (~6 MB) lần đầu chạy.
# Nếu không muốn cài ultralytics, đổi DETECTOR_BACKEND = "hog" trong config.py

flask>=2.3.0
opencv-contrib-python>=4.5.0
numpy>=1.21.0
qrcode[pil]>=7.3.1
Pillow>=9.0.0
ultralytics>=8.0.0
pyserial>=3.5

# Xác minh khuôn mặt (tùy chọn — cần cmake + build-essential):
# face_recognition>=1.3.0

# ── Raspberry Pi only ────────────────────────────────────────────────────────
# Thêm khi chạy trên RPi (HARDWARE_MODE = "raspi"):
#   pip install RPi.GPIO
#
# Trên RPi, dùng opencv-python thay vì opencv-contrib-python:
#   pip install opencv-python>=4.5.0
#
# Bật UART:
#   sudo raspi-config → Interface Options → Serial Port
#     "login shell over serial?" → No
#     "serial port hardware enabled?" → Yes
#   sudo reboot
